+++ hive_generator/lib/src/builder.dart
@@ -1,3 +1,4 @@
+import 'package:analyzer/dart/constant/value.dart';
 import 'package:analyzer/dart/element/element.dart';
 import 'package:analyzer/dart/element/type.dart';
 
@@ -5,7 +6,7 @@ class AdapterField {
   final int index;
   final String name;
   final DartType type;
-  final bool defaultValue;
+  final DartObject? defaultValue;
 
   AdapterField(this.index, this.name, this.type, this.defaultValue);
 }